reorganise structure of app
