+++ src/main/java/com/pip/dao/AmmunitionDao.java
@@ -1,10 +1,9 @@
 package dao;
 
-import entities.Ammunition;
+import com.pip.entities.Ammunition;
+import org.springframework.stereotype.Repository;
 
-import javax.ejb.Stateless;
-
-@Stateless
+@Repository
 public class AmmunitionDao extends CommonDao<Ammunition> {
     public AmmunitionDao() {
         super(Ammunition.class);
