+++ src/main/java/com/pip/dao/EngineDao.java
@@ -1,10 +1,9 @@
 package dao;
 
-import entities.Engine;
+import com.pip.entities.Engine;
+import org.springframework.stereotype.Repository;
 
-import javax.ejb.Stateless;
-
-@Stateless
+@Repository
 public class EngineDao extends CommonDao<Engine> {
     public EngineDao() {
         super(Engine.class);
