+++ src/main/java/com/pip/dao/FirmWeaponDao.java
@@ -1,10 +1,9 @@
 package dao;
 
-import entities.FirmWeapon;
+import com.pip.entities.FirmWeapon;
+import org.springframework.stereotype.Repository;
 
-import javax.ejb.Stateless;
-
-@Stateless
+@Repository
 public class FirmWeaponDao extends CommonDao<FirmWeapon> {
     public FirmWeaponDao() {
         super(FirmWeapon.class);
