+++ src/main/java/com/pip/entities/SpecializationHuman.java
