+++ src/main/java/com/pip/enums/TypeFuel.java
