+++ src/main/java/com/pip/services/HWService.java
@@ -2,7 +2,7 @@ package services;
 
 import javax.enterprise.context.Dependent;
 
-@Dependent
+//@Dependent
 public class HWService {
 
     public String createHelloMessage(String name) {
