--- src/main/webapp/index.jsp
@@ -1,16 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: ritt
-  Date: 10/6/18
-  Time: 2:05 PM
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page import="java.util.*" contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-  <head>
-    <title>third</title>
-  </head>
-  <body>
-  <% System.out.print("rita"); %>
-  </body>
-</html>
